+++ packages/expo-router/CHANGELOG.md
@@ -14,6 +14,7 @@
 
 - Remove error hiding system. ([#26607](https://github.com/expo/expo/pull/26607) by [@EvanBacon](https://github.com/EvanBacon))
 - Make `@testing-library/jest-native` usage optional ([#26650](https://github.com/expo/expo/pull/26650) by [@marklawlor](https://github.com/marklawlor))
+- Include search parameters in the default Screen.getId() function.
 
 ### 💡 Others